requirements.txt
pydantic == 2.7.4
bittensor == 7.4.0
polygon-api-client == 1.13.7
scikit-learn == 1.5.0
scipy == 1.12.0
pandas == 2.2.2
twelvedata == 1.2.24
flask == 3.0.3
waitress == 2.1.2
matplotlib == 3.9.0
sortedcontainers == 2.4.0
pandas-market-calendars == 4.4.0
google-cloud-storage == 2.17.0
python-dotenv
alpaca-trade-api

